chore: start with smoke test for dashboard chore: improve TranslationComponent required logic fix: a bug with TodaysOrders count chore: remove ViewAdvertisement because it is not used chore: improve AdvertisementForm and fix some bugs fix: a bug with ProductForm branch config values being decimal chore: improve ProductInfolist design chore: improve AdvertisementForm preview design and fix some bugs chore: add permission check for /docs route

+++ app/Filament/Resources/Products/Schemas/ProductForm.php
@@ -113,12 +113,12 @@ class ProductForm
                     TextInput::make('minimum_order_quantity')
                         ->label('Min Order Qty')
                         ->required()
-                        ->integer()
+                        ->numeric()
                         ->rule('gt:0'),
                     TextInput::make('maximum_order_quantity')
                         ->label('Max Order Qty')
                         ->required()
-                        ->integer()
+                        ->numeric()
                         ->rule(function (Get $get) {
                             $minimumOrderQuantity = $get('minimum_order_quantity');
                             if ($minimumOrderQuantity === null || $minimumOrderQuantity < 1) {
